feat(stream-text): introduce CompletionToolCall and CompletionToolResult interfaces
